requirements/req_dev.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile req_dev.in
#
alabaster==0.7.16
    # via sphinx
amqp==5.2.0
    # via kombu
asgiref==3.8.1
    # via django
attrs==23.2.0
    # via
    #   outcome
    #   trio
babel==2.14.0
    # via sphinx
billiard==4.2.0
    # via celery
celery==5.4.0
    # via
    #   -r req_dev.in
    #   django-celery-results
certifi==2024.2.2
    # via
    #   requests
    #   selenium
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
coverage[toml]==7.4.4
    # via pytest-cov
django==5.0.4
    # via
    #   -r req_dev.in
    #   django-appconf
    #   django-braces
    #   django-celery-results
    #   django-debug-toolbar
    #   django-filter
    #   djangorestframework
django-appconf==1.0.6
    # via django-compressor
django-braces==1.15.0
    # via -r req_dev.in
django-celery-results==2.5.1
    # via -r req_dev.in
django-compressor==4.4
    # via -r req_dev.in
django-debug-toolbar==4.3.0
    # via -r req_dev.in
django-discover-runner==1.0
    # via -r req_dev.in
django-filter==24.2
    # via -r req_dev.in
djangorestframework==3.15.1
    # via -r req_dev.in
docutils==0.21.1
    # via sphinx
ebooklib==0.18
    # via -r req_dev.in
factory-boy==3.3.0
    # via -r req_dev.in
faker==24.11.0
    # via
    #   -r req_dev.in
    #   factory-boy
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via -r req_dev.in
h11==0.14.0
    # via wsproto
idna==3.7
    # via
    #   requests
    #   trio
imagesize==1.4.1
    # via sphinx
iniconfig==2.0.0
    # via pytest
jinja2==3.1.3
    # via sphinx
kombu==5.3.7
    # via celery
lxml==5.2.1
    # via
    #   -r req_dev.in
    #   ebooklib
    #   python-ooxml
markdown==3.6
    # via -r req_dev.in
markupsafe==2.1.5
    # via jinja2
mock==5.1.0
    # via -r req_dev.in
outcome==1.3.0.post0
    # via trio
packaging==24.0
    # via
    #   pytest
    #   sphinx
pep8==1.7.1
    # via -r req_dev.in
pika==1.3.2
    # via -r req_dev.in
pillow==10.3.0
    # via -r req_dev.in
pluggy==1.5.0
    # via pytest
prompt-toolkit==3.0.43
    # via click-repl
psycopg2-binary==2.9.9
    # via -r req_dev.in
pygments==2.17.2
    # via sphinx
pysocks==1.7.1
    # via urllib3
pytest==8.1.1
    # via
    #   -r req_dev.in
    #   pytest-cov
    #   pytest-django
    #   pytest-pythonpath
pytest-cov==5.0.0
    # via -r req_dev.in
pytest-django==4.8.0
    # via -r req_dev.in
pytest-pythonpath==0.7.3
    # via -r req_dev.in
python-dateutil==2.9.0.post0
    # via
    #   celery
    #   faker
python-ooxml==0.13
    # via -r req_dev.in
rcssmin==1.1.1
    # via django-compressor
redis==5.0.3
    # via -r req_dev.in
requests==2.31.0
    # via
    #   -r req_dev.in
    #   sphinx
rjsmin==1.2.1
    # via django-compressor
selenium==4.19.0
    # via -r req_dev.in
six==1.16.0
    # via
    #   ebooklib
    #   python-dateutil
    #   python-ooxml
smmap==5.0.1
    # via gitdb
sniffio==1.3.1
    # via trio
snowballstemmer==2.2.0
    # via sphinx
sortedcontainers==2.4.0
    # via trio
sphinx==7.3.7
    # via -r req_dev.in
sphinxcontrib-applehelp==1.0.8
    # via sphinx
sphinxcontrib-devhelp==1.0.6
    # via sphinx
sphinxcontrib-htmlhelp==2.0.5
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.7
    # via sphinx
sphinxcontrib-serializinghtml==1.1.10
    # via sphinx
sqlparse==0.5.0
    # via
    #   -r req_dev.in
    #   django
    #   django-debug-toolbar
trio==0.25.0
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.11.1
    # via selenium
typing-extensions==4.11.0
    # via selenium
tzdata==2024.1
    # via celery
unidecode==1.3.8
    # via -r req_dev.in
unipath==1.1
    # via -r req_dev.in
urllib3[socks]==2.2.1
    # via
    #   requests
    #   selenium
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
wsproto==1.2.0
    # via trio-websocket
